Convert to webpack + vue-loader way - must run this up, repair errors
