+++ view/frontend/templates/onepage.phtml
@@ -1,232 +1,6 @@
 <?= $this->getChildHtml() ?>
 
-<div id="checkout" class="checkout-container">
-    <section class="shipping-information" v-if="step === 'addresses'">
-        <h1>
-            <?= /* @escapeNotVerified */ __('Billing Address'); ?>
-        </h1>
-        
-        <form class="billing-address__form">
-            <template v-for="field in billingAddress">
-                <template v-if="field.type !== 'select'">
-                    <inputfield 
-                        :label="field.label" 
-                        :name="field.name" 
-                        :type="field.type"
-                        :value="field.value"
-                        fieldclass="billing-address__field"
-                        inputclass="input billing-address__input"
-                    />
-                </template>
-                
-                <template v-if="field.type === 'select'">
-                    <selectfield
-                        :label="field.label"
-                        :name="field.name"
-                        :options="field.options"
-                        fieldclass="billing-address__field"
-                        selectclass="billing-address__select"
-                        @change.native="changeSelection"
-                    />
-                </template>
-
-                <template v-if="field.name === 'region_id'">
-                    <inputfield 
-                        label="State/Province" 
-                        name="region" 
-                        type="text"
-                        fieldclass="billing-address__field region--hidden"
-                        inputclass="input billing-address__input"
-                    />
-                </template>
-            </template>
-        </form>
-
-        <h1>
-            <?= /* @escapeNotVerified */ __('Shipping address'); ?>
-        </h1>
-
-        <template>
-            <checkboxfield 
-                id="shippingAddress"
-                labelclass="label"
-                fieldclass="checkbox shipping-address__field"
-                inputclass="shipping-address__checkbox"
-                checked="true"
-                name="shippingAddress"
-                text="My billing and shipping address are the same"
-                @change.native="toggleShippingAddress"
-            />
-        </template>
-
-        <form class="shipping-address__form shipping-address--hidden">
-            <template v-for="field in shippingAddress">
-                <template v-if="field.type !== 'select'">
-                    <inputfield 
-                        :label="field.label" 
-                        :name="field.name" 
-                        :type="field.type"
-                        :value="field.value"
-                        fieldclass="shipping-address__field"
-                        inputclass="input shipping-address__input"
-                    />
-                </template>
-
-                <template v-if="field.type === 'select'">
-                    <selectfield
-                        :label="field.label"
-                        :name="field.name"
-                        :options="field.options"
-                        fieldclass="shipping-address__field"
-                        selectclass="shipping-address__select"
-                        @change.native="changeSelection"
-                    />
-                </template>
-
-                <template v-if="field.name === 'region_id'">
-                    <inputfield 
-                        label="State/Province" 
-                        name="region" 
-                        type="text"
-                        fieldclass="shipping-address__field region--hidden"
-                        inputclass="input shipping-address__input"
-                    />
-                </template>
-            </template>
-
-            <template>
-                <buttonelement
-                    buttontype="button"
-                    buttonclass="button"
-                    text="Cancel"
-                    @click.native="cancelShippingInformations"
-                />
-            </template>
-        </form>
-
-        <template>
-            <buttonelement
-                buttontype="button"
-                buttonclass="button"
-                text="Set Shipping Informations"
-                @click.native="setShippingInformation"
-            />
-        </template>
-    </section>
-
-    <section class="methods" v-if="step === 'methods'">
-        <h1>
-            <?= /* @escapeNotVerified */ __('Choose shipping and payment methods'); ?>
-        </h1>
-
-        <h2>
-            <?= /* @escapeNotVerified */ __('Shipping methods'); ?>
-        </h2>
-
-        <template>
-            <shipping-methods
-                :options="shippingMethods"
-                :currencycode="totals.base_currency_code"
-                name="shipping"
-                labelclass="labels"
-                containerclass="methods__handler"
-                fieldclass="radio methods__field"
-                inputclass="methods__radio"
-            />
-        </template>
-
-        <h2>
-            <?= /* @escapeNotVerified */ __('Payment methods'); ?>
-        </h2>
-
-        <template>
-            <payment-methods
-                :options="paymentMethods"
-                name="payment"
-                labelclass="labels"
-                containerclass="methods__handler"
-                fieldclass="radio methods__field"
-                inputclass="methods__radio"
-            />
-        </template>
-
-        <template>
-            <buttonelement
-                buttontype="button"
-                buttonclass="button"
-                text="Go to summary"
-                @click.native="setMethods"
-            />
-        </template>
-
-        <template>
-            <buttonelement
-                buttontype="button"
-                buttonclass="button"
-                text="Back"
-                @click.native="step = 'summary'"
-            />
-        </template>
-    </section>
-
-    <section v-if="step === 'summary'">
-        <h1>
-            <?= /* @escapeNotVerified */ __('Summary'); ?>
-        </h1>
-
-        <template>
-            <summation
-                :items="totals.total_segments"
-                :discountamout="totals.discount_amount"
-                :itemsqty="totals.items_qty"
-                :currencycode="totals.base_currency_code"
-            />
-        </template>
-
-        <template>
-            <buttonelement
-                buttontype="button"
-                buttonclass="button"
-                text="Place order"
-                @click.native="makeOrder"
-            />
-        </template>
-        
-        <template>
-            <buttonelement
-                buttontype="button"
-                buttonclass="button"
-                text="Back"
-                @click.native="step = 'methods'"
-            />
-        </template>
-    </section>
-
-    <section v-if="step === 'success'">
-        <template>
-            <?= /* @escapeNotVerified */ __('We did it!'); ?>
-
-            <a href="/what-is-new.html">
-                <?= /* @escapeNotVerified */ __('Back to category'); ?>
-            </a>
-        </template>
-    </section>
-
-    <section v-if="step !== 'success'">
-        <h1>
-            <?= /* @escapeNotVerified */ __('Product List'); ?>
-        </h1>
-
-        <template>
-            <products 
-                containerclass="grid"
-                itemclass="grid__columns"
-                :products="config.totalsData.items"
-                :currency="config.totalsData.base_currency_code"
-            />
-        </template>
-    </section>
-</div>
+<div id="checkout" class="checkout-container"></div>
 
 <style>
 .region--hidden,
