+++ view/frontend/web/js/components/payment-methods.vue
@@ -1,0 +1,43 @@
+<template>
+    <div :class="containerclass">
+        <template v-for="option in options">
+            <div :class="fieldclass">
+                    <input type="radio" 
+                            :id="option.code" 
+                            :name="name" 
+                            :value="option.code"
+                            :class="inputclass" 
+                    />
+
+                    <label :class="labelclass" :for="option.code">
+                        {{ option.title }}
+                    </label>
+            </div>
+        </template>
+    </div>
+</template>
+
+<script>
+export default {
+    props: {
+        options: {
+            type: Array
+        },
+        labelclass: {
+            type: String
+        },
+        name: {
+            type: String
+        },
+        containerclass: {
+            type: String
+        },
+        fieldclass: {
+            type: String
+        },
+        inputclass: {
+            type: String
+        }
+    }
+}
+</script>
