--- view/frontend/web/js/filters/currency.js
@@ -1,9 +1,0 @@
-define([
-    'vue'
-], function(Vue) {
-    'use strict'
-   
-    Vue.filter('currency', function(value) {
-        return parseFloat(value).toFixed(2);
-    });
-});
