--- view/frontend/web/js/filters/trimZero.js
@@ -1,9 +1,0 @@
-define([
-    'vue'
-], function(Vue) {
-    'use strict'
-   
-    Vue.filter('trimZero', function(value) {
-        return parseInt(value);
-    });
-});
